Remove sample app package hacks and clearly separate target and test apps
The way instrumented tests work is that the tests are in a test APK,
and the app is in a host APK. The test APK is injected into the host's
process, and the tests run there.

When Gradle builds the test apk, it bakes into the manifest the
package of the host app, which is normally the target of the tests,
as "targetPackage".

In our case, we want to run instrumented tests against other apps.
So far, we hacked into the Gradle build to replace the ID of the host
app with the ID of the app that we intended to test. This way, the
test APK's manifest actually declared that it needed to be injected
into our app. We never installed the actual host app that was part
of the Gradle project of the instrumented tests. Instead, we
installed the app under test which happened to match the ID.

Now that forces us to rebuild the instrumented test app for each test,
and is also quite hacky (and confusing).

The instrumented tests actually run 'monkey -p $targetAppPackage' so
they don't really need to live in the process of the app under test.
Therefore, we can keep the original host APK of the instrumented tests
to host the tests, and the tests can still call our app under test via
monkey. We just need to make sure we install the host APK now.