Update IJ Platform to 261.26222.65: jvm-inc-builder: use diagnostic.toString() to render the compiler message as a String: AMPER-5436
The previous implementation tried to construct a String using the structured information about the compiler message. But sometimes messages, which are reported against a class declaration, indicate the whole class body as the source of the problem. Therefore, posting the code between `diagnostic.getStartPosition()` and `diagnostic.getEndPosition()` means posting the whole class body to the compiler output.

Ideally, the output should be sent in a structured format, and the caller should deal with the information in the most convenient way for the given UI. But until that is implemented, let's use the default `diagnostic.toString()` implementation that is good enough.